NewArmor-1

+++ Assets/Scripts/Farm.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Farm : MonoBehaviour
 {
     public Nodes nodes;
-    public int collectAmount = 0;
+    public float collectAmount = 0;
     public int MaxFarmWorker = 1;
     public int currentWorkerCount = 0;
 }
